Remove uploadableTextures from fullProfile
As far as I'm aware, authlib-injector does not use this anywhere. I'd
like if it were removed because in ReIndev, a popular b1.7.3 mod, it
attempts to base64 decode every "value" field in `properties`. Under
Mojang authentication server it succeeds, but with drasl it will try to
Base64-decode something like "skin,cape" and fail.